# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/intents": {
            "get": {
                "description": "按 kind、status、resource、application 等维度查询 execution intents",
                "tags": [
                    "Intent"
                ],
                "summary": "获取执行意图列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Intent"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/intents/{id}": {
            "get": {
                "tags": [
                    "Intent"
                ],
                "summary": "获取执行意图",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Intent ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Intent"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/jobs": {
            "get": {
                "tags": [
                    "Job"
                ],
                "summary": "获取Job列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Job"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "创建一个新的Job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Job"
                ],
                "summary": "创建Job",
                "parameters": [
                    {
                        "description": "Job Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Job"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg_api.CreateResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/jobs/{id}": {
            "get": {
                "tags": [
                    "Job"
                ],
                "summary": "获取Job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Job"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Job"
                ],
                "summary": "更新Job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Job Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.Job"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Job"
                ],
                "summary": "删除Job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_bsonger_devflow-app-service_pkg_model.Intent": {
            "type": "object",
            "properties": {
                "application_id": {
                    "type": "string"
                },
                "application_name": {
                    "type": "string"
                },
                "attempt_count": {
                    "type": "integer"
                },
                "branch": {
                    "type": "string"
                },
                "claimed_at": {
                    "type": "string"
                },
                "claimed_by": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "env": {
                    "type": "string"
                },
                "external_ref": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "job_id": {
                    "type": "string"
                },
                "job_type": {
                    "type": "string"
                },
                "kind": {
                    "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.IntentKind"
                },
                "last_error": {
                    "type": "string"
                },
                "lease_expires_at": {
                    "type": "string"
                },
                "manifest_id": {
                    "type": "string"
                },
                "manifest_name": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "repo_url": {
                    "type": "string"
                },
                "resource_id": {
                    "type": "string"
                },
                "resource_type": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.IntentStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_bsonger_devflow-app-service_pkg_model.IntentKind": {
            "type": "string",
            "enum": [
                "build",
                "release"
            ],
            "x-enum-varnames": [
                "IntentKindBuild",
                "IntentKindRelease"
            ]
        },
        "github_com_bsonger_devflow-app-service_pkg_model.IntentStatus": {
            "type": "string",
            "enum": [
                "Pending",
                "Running",
                "Succeeded",
                "Failed"
            ],
            "x-enum-varnames": [
                "IntentPending",
                "IntentRunning",
                "IntentSucceeded",
                "IntentFailed"
            ]
        },
...
        "github_com_bsonger_devflow-app-service_pkg_model.Job": {
            "type": "object",
            "properties": {
                "application_id": {
                    "type": "string"
                },
                "application_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "env": {
                    "type": "string"
                },
                "execution_intent_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "manifest_id": {
                    "type": "string"
                },
                "manifest_name": {
                    "type": "string"
                },
                "project_name": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.JobStatus"
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.JobStep"
                    }
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_bsonger_devflow-app-service_pkg_model.JobStatus": {
            "type": "string",
            "enum": [
                "Pending",
                "Running",
                "Succeeded",
                "Failed",
                "RollingBack",
                "RolledBack",
                "Syncing",
                "SyncFailed"
            ],
            "x-enum-varnames": [
                "JobPending",
                "JobRunning",
                "JobSucceeded",
                "JobFailed",
                "JobRollingBack",
                "JobRolledBack",
                "JobSyncing",
                "JobSyncFailed"
            ]
        },
        "github_com_bsonger_devflow-app-service_pkg_model.JobStep": {
            "type": "object",
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "progress": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_bsonger_devflow-app-service_pkg_model.StepStatus"
                }
            }
        },
...
        "github_com_bsonger_devflow-app-service_pkg_model.StepStatus": {
            "type": "string",
            "enum": [
                "Pending",
                "Running",
                "Succeeded",
                "Failed"
            ],
            "x-enum-varnames": [
                "StepPending",
                "StepRunning",
                "StepSucceeded",
                "StepFailed"
            ]
        },
        "pkg_api.CreateResponse": {
            "type": "object",
            "properties": {
                "execution_intent_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
    }
}